server/requirements.txt
# Chess Game Server Requirements - Production Ready
# WebSocket server
websockets==12.0
uvicorn[standard]==0.24.0

# Environment management
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0

# Game engine dependencies
opencv-python==4.8.1.78
numpy==1.24.3

# Optional: For HTTP health checks
fastapi==0.104.1
